ДЗ 2: рефакторинг
