+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:supportsRtl="true"
             android:theme="@style/Theme.BaForecast.NoActionBar">
         <activity
-                android:name=".MainActivity"
+                android:name=".activity.MainActivity"
                 android:label="@string/app_name"
                 android:theme="@style/Theme.BaForecast.NoActionBar">
             <intent-filter>
@@ -22,11 +22,11 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".SelectCityActivity"
+                android:name=".activity.SelectCityActivity"
                 android:label="@string/select_city_name"
                 android:theme="@style/Theme.BaForecast.NoActionBar" />
         <activity
-                android:name=".SettingsActivity"
+                android:name=".activity.SettingsActivity"
                 android:label="@string/settings_name"
                 android:theme="@style/Theme.BaForecast.NoActionBar" />
     </application>
