+++ app/src/main/java/com/example/baforecast/activity/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.example.baforecast;
+package com.example.baforecast.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,6 +10,7 @@ import android.widget.Toast;
 import androidx.appcompat.widget.SwitchCompat;
 import androidx.appcompat.widget.Toolbar;
 
+import com.example.baforecast.R;
 import com.google.android.material.textfield.TextInputEditText;
 
 public class SettingsActivity extends BaseActivity {
