+++ app/src/main/java/com/example/baforecast/adapter/SocnetAdapter.java
@@ -1,4 +1,4 @@
-package com.example.baforecast;
+package com.example.baforecast.adapter;
 
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.baforecast.R;
 
 public class SocnetAdapter extends RecyclerView.Adapter<SocnetAdapter.ViewHolder> {
     private final String[] days;
