+++ app/src/main/java/com/example/baforecast/constant/Constants.java
@@ -1,7 +1,8 @@
-package com.example.baforecast;
+package com.example.baforecast.constant;
 
 public class Constants {
     public static final String EXTRA_PARCEL = "parcel";
     public static final String EXTRA_CURRENT_CITY = "CurrentCity";
     public static final int SETTING_CODE = 88;
+    public static final String WEATHER_URL = "https://api.openweathermap.org/data/2.5/weather?q=%s,RU&appid=%s";
 }
