+++ app/src/main/java/com/example/baforecast/observer/Observer.java
@@ -1,6 +1,6 @@
 package com.example.baforecast.observer;
 
-import com.example.baforecast.City;
+import com.example.baforecast.model.City;
 
 public interface Observer {
     void selectCity(City cityParcel);
