+++ app/src/main/java/com/example/baforecast/source/WeatherDisplayable.java
@@ -1,0 +1,7 @@
+package com.example.baforecast.source;
+
+import com.example.baforecast.model.WeatherRequest;
+
+public interface WeatherDisplayable {
+    void displayWeather(WeatherRequest request);
+}
