+++ app/src/main/res/layout/activity_base.xml
@@ -4,6 +4,6 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".BaseActivity">
+        tools:context=".activity.BaseActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
