+++ app/src/main/res/layout/app_bar_select_city.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SelectCityActivity">
+        tools:context=".activity.SelectCityActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
