+++ app/src/main/res/layout/fragment_cities.xml
@@ -4,6 +4,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".CitiesFragment">
+        tools:context=".fragment.CitiesFragment">
 
 </LinearLayout>
