+++ app/src/main/res/menu/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".MainActivity"
+    tools:context=".activity.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
     <ScrollView
@@ -31,7 +31,7 @@
                         android:id="@+id/cityInfo"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
-                        android:name="com.example.baforecast.CityInfoFragment" />
+                        android:name="com.example.baforecast.fragment.CityInfoFragment" />
             </FrameLayout>
 
             <androidx.recyclerview.widget.RecyclerView
